Move to filter decorator ot generic
